[Remote Config] Add exponential backoff to HTTP client
Adds exponential backoff to the remote config HTTP client to
prevent agents from overwhelming the remote config edge during service
delivery issues.

To do this, we leverage pkg/util/backoff and define a `backoff.Policy`
for the remote config client. We keep track of the current error level
on successful and failed updates based on the defined `Policy`.

The backoff policy has jitter built in for random backoff times. The
default values we use for the policy start with a backoff duration that
will be within the range [1s, 2s], with the min and max doubling with
every consecutive error. When we see a success, we drop back two states.
(or the current interval / 4). The maximum backoff added is
configurable, and defaults to 5m. No other values are configurable by
the client at this time.

Resolves: [RCM-58]